+++ src/app/store/index.ts
@@ -1,1 +1,4 @@
+export * from './model/types';
+export * from './useAppStore';
+
 export { default as store } from './model/store';
