+++ src/features/theme/index.ts
@@ -1,3 +1,4 @@
 export * from './model/themeSlice';
 
 export { default as ThemeWrapper } from './ui/ThemeWrapper';
+export { default as ModeSwitcher } from './ui/ModeSwitcher';
