--- src/modules/theme/hooks/ui/useAppTheme.ts
@@ -1,13 +1,0 @@
-import { useMemo } from 'react';
-import { useAppSelector } from '@/app/store/useAppStore';
-import { selectThemeMode } from '@/features/theme';
-import { Theme } from '@mui/material';
-import { createAppTheme } from '@/shared/theme';
-
-const useAppTheme = (): Theme => {
-  const mode = useAppSelector(selectThemeMode);
-
-  return useMemo(() => createAppTheme(mode), [mode]);
-};
-
-export default useAppTheme;
