--- src/modules/theme/index.ts
@@ -1,2 +1,0 @@
-export { default as ThemeWrapper } from './ui/containers/ThemeWrapper';
-export { default as ModeSwitcher } from './ui/containers/ModeSwitcher';
