--- src/modules/theme/ui/containers/ThemeWrapper.tsx
@@ -1,20 +1,0 @@
-import { CssBaseline, ThemeProvider } from '@mui/material';
-import { ReactNode } from 'react';
-import useAppTheme from '@/features/theme/model/useAppTheme';
-
-interface ThemeWrapperProps {
-  children: ReactNode;
-}
-
-const ThemeWrapper = ({ children }: ThemeWrapperProps) => {
-  const theme = useAppTheme();
-
-  return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      {children}
-    </ThemeProvider>
-  );
-};
-
-export default ThemeWrapper;
