+++ src/shared/theme/config/getPalette.ts
@@ -18,7 +18,7 @@ const avatarColors = {
   green: green[600],
 } as const;
 
-type AvatarColors = typeof avatarColors;
+export type AvatarColors = typeof avatarColors;
 type CustomBackground = {
   avatar: AvatarColors;
   userCard: string;
@@ -44,5 +44,5 @@ const getPalette = (mode: ThemeMode): PaletteOptions => {
   };
 };
 
-export type AvatarColor = AvatarColors[keyof AvatarColors];
 export default getPalette;
+export type AvatarColor = AvatarColors[keyof AvatarColors];
