+++ src/shared/theme/index.ts
@@ -1,4 +1,6 @@
 export * from './model/types';
 export * from './model/storageHandlers';
 
+export type { AvatarColor } from './config/getPalette';
+
 export { default as createAppTheme } from './config/createAppTheme';
