+++ src/shared/theme/model/types.ts
@@ -2,5 +2,3 @@ export enum ThemeMode {
   Dark = 'dark',
   Light = 'light',
 }
-
-export type { AvatarColor } from '../config/getPalette';
