--- src/store/index.ts
@@ -1,4 +1,0 @@
-export * from './model/useAppStore';
-export * from './model/store';
-
-export { default as store } from './model/store';
