--- src/store/model/root-reducer.ts
@@ -1,8 +1,0 @@
-import { combineSlices } from '@reduxjs/toolkit';
-import { themeSlice } from '@/features/theme';
-import { usersSlice } from '@/modules';
-import { baseAPI } from '@/shared/api';
-
-const rootReducer = combineSlices(baseAPI, themeSlice, usersSlice);
-
-export default rootReducer;
