--- src/store/model/store.ts
@@ -1,11 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { baseAPI } from '@/shared/api';
-import rootReducer from './root-reducer';
-
-const store = configureStore({
-  reducer: rootReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(baseAPI.middleware),
-});
-
-export default store;
