--- src/store/model/types.ts
@@ -1,11 +1,0 @@
-import { ThunkAction, UnknownAction } from '@reduxjs/toolkit';
-import store from './store';
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-export type AppThunk<ReturnType = void> = ThunkAction<
-  ReturnType,
-  RootState,
-  unknown,
-  UnknownAction
->;
