--- src/store/model/useAppStore.tsx
@@ -1,5 +1,0 @@
-import { useDispatch, useSelector } from 'react-redux';
-import { AppDispatch, RootState } from './types';
-
-export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
-export const useAppSelector = useSelector.withTypes<RootState>();
